pyproject.toml
[tool.poetry]
name = "django-potencia-tech"
version = "0.1.0"
description = ""
authors = ["Ícaro Abreu <136408366+oicaroabreu@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.2"
django = "^4.2.6"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
